refactor(json): migrate Gson → kotlinx.serialization
Replace Gson across persistence and the Nostr wire layer with
kotlinx.serialization, introducing a shared JsonConfig (omit nulls,
ignoreUnknownKeys, lenient) that mirrors the previous Gson behaviour so
existing persisted data and relay/peer JSON stay compatible.

- Nostr: NostrEvent/Filter/Request/Protocol/RelayManager now use
  kotlinx custom serializers. NIP-01 event-id hashing keeps its exact
  compact byte layout via buildJsonArray (no slash escaping), preserving
  event IDs and signatures.
- Persistence: FavoritesPersistenceService, SeenMessageStore,
  GeohashBookmarksStore, LocationChannelManager, DataManager,
  NoiseChannelEncryption, OpenStreetMapGeocoderProvider.
- BinaryProtocol: replace MAX_PAYLOAD_LENGTH.toUInt() comparisons with
  toLong() to dodge a Kotlin IR const-eval back-end crash triggered once
  the serialization plugin enables const evaluation. Wire format
  unchanged.
- Build: add kotlin.plugin.serialization, drop gson; catalog gson →
  kotlinx-serialization-json.

+++ app/src/main/java/com/bitchat/android/protocol/BinaryProtocol.kt
@@ -365,7 +365,7 @@ object BinaryProtocol {
                 buffer.getShort().toUShort().toUInt()  // 2 bytes for v1, convert to UInt
             }
 
-            if (payloadLength > com.bitchat.android.util.AppConstants.Protocol.MAX_PAYLOAD_LENGTH.toUInt()) {
+            if (payloadLength.toLong() > com.bitchat.android.util.AppConstants.Protocol.MAX_PAYLOAD_LENGTH.toLong()) {
                 Log.w("BinaryProtocol", "Payload length ${payloadLength} exceeds maximum allowed (${com.bitchat.android.util.AppConstants.Protocol.MAX_PAYLOAD_LENGTH})")
                 return null
             }
